pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*"]

[project]
name = "rag-new"
version = "0.1.0"
description = "RAG: local Chroma by default (rag/main.ipynb style); optional Milvus hybrid + BM25 when VECTOR_BACKEND=milvus"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "python-multipart>=0.0.9",
    "uvicorn>=0.30.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.0",
    "pymilvus>=2.5.0",
    "pydantic>=2.8.0",
    "langchain>=0.2.14",
    "langchain-core>=0.2.37",
    "langchain-community>=0.2.12",
    "langchain-text-splitters>=0.2.2",
    "langchain-huggingface>=0.1.0",
    "langchain-openai>=0.1.22",
    "sentence-transformers>=3.0.0",
    "langgraph>=0.2.31",
    "sqlalchemy>=2.0.36",
    "pypdf>=4.3.1",
    "docx2txt>=0.8",
    "unstructured",
    "openpyxl",
    "chromadb>=0.5.0",
]
